generalize EOIE exstension writing
